Add scroll and fix UI
Issue #47

+++ lib/common_models/category_model.dart
@@ -11,9 +11,9 @@ class Category {
 
   Category.fromJson(Json categoryJson) {  // TODO- extract strings to firebase_config (also from user_model and transaction)
     name = categoryJson["name"];
-    expected = categoryJson["expected"];
+    expected = categoryJson["expected"].toDouble();
     description = categoryJson["description"];
-    amount = categoryJson["amount"];
+    amount = categoryJson["amount"].toDouble();
     isIncome = categoryJson["isIncome"];
     transactions = jsonToElementList(categoryJson["transactions"], (json) => Transaction.fromJson(json)).cast<Transaction>();
   }
